+++ compile_all.sh
@@ -1,2 +1,0 @@
-# compile gsf
-gfortran -o gsf/a.out gsf/spher_expan.f