+++ src/scripts/generate_geos_optics.legacy.sh
@@ -1,3 +1,12 @@
+#!/bin/tcsh
+
+# setup environment
+setenv SRC_DIR @SRCDIR
+setenv PYTHONPATH ${SRC_DIR}/install/lib/Python
+
+source $SRC_DIR/env@/g5_modules
+
+
 # full suite assumes geosparticles/du.json has a valid path to dust kernels
 # which are not included by default in GEOSmie distribution
 #./runoptics.py --namelist namelists/geos.txt